refactor(createTask): separated page and component structures
