+++ src/App.tsx
@@ -9,7 +9,7 @@ import Employees from './pages/Employees';
 import Home from './pages/Home';
 import Projects from './pages/Projects';
 import Tasks from './pages/Tasks';
-import NewTask from './pages/Tasks/new';
+import NewTaskPage from './pages/Tasks/new';
 import { RoutesPath } from './utils/constants';
 
 function App() {
@@ -61,14 +61,7 @@ function App() {
             />
           </Route>
 
-          <Route
-            path={`/tasks/create`}
-            element={
-              <ProtectedRoute>
-                <NewTask />
-              </ProtectedRoute>
-            }
-          />
+          <Route path={'/tasks/create'} element={<NewTaskPage />} />
         </Routes>
       </Router>
     </LocalizationProvider>
