+++ src/types/task.ts
@@ -1,15 +1,12 @@
 import { TaskStatus } from './task-status';
 
-export interface TaskEntity {
-  id: string;
+export interface BareboneTask {
   title: string;
   description: string;
   status: TaskStatus;
-  waitingFor?: string;
-  startDate: Date | null;
-  endDate?: Date | null;
-  workedHours?: number;
-  createdAt: Date;
-  updatedAt: Date | null;
+  waitingFor: string;
+  startDate: string | null;
+  dueDate: string | null;
+  workedHours: string | null;
   idProject: string;
 }
